Gah, finally get a sensible combination of rsync options
Note that, since we're using Nix, all timestamps on the source will be
1970-01-01 and hence all of the new files will appear to be older than the
previous set.

This combination of options will calculate checksums of the local files, send
that over to the other end, which will compare them, and ask for any mis-matched
files to be transferred.

This can take a couple of minutes, but it seems to a) cause updates to propagate
and b) doesn't send anything when run multiple times (i.e. it's idempotent)