Don't rely on out-of-memory for tail-recursive unique
We were exhausting the system's memory to prove a point about non-tail-recursive
functions being inefficient; but that just makes the system crawl and takes
ages. Much better to simply measure memory usage and check if some limit is
reached; that way, we can set the limit much lower and not kill the system.