Remove some git-related env vars
These no longer seem to be used. They're presumably left over from when we
rendered each git repo's pages as part of this site; that's since been moved to
a separate tool (in warbo-utilities), so this repo no longer renders the
contents of /git. Our /projects/repos entries are simple redirects to /git.