Add htmlUnwrap command
This is like panhandle, but used as a post-processor for the HTML. It unwraps
the content of <pre class="htmlunwrap"><code>...</code></pre>

I made this to avoid problems embedding HTML in markdown, in particular iframes.
We could actually use panhandle for this, but the way pandoc was handling HTML
in my experiments seemed a bit off; so I made a quick Python script using
BeautifulSoup instead.