sk in egglog: don't artificially limit Fuel
If a generator takes a Fuel argument, pass that along to calls which need it,
rather than using genFuelN to choose some Fuel *up to* that amount.

Likewise, if a Range is needed, pass the whole Range we've been given; not some
Range *up to* a value chosen from what we're given.

This causes falsify to work harder, generating more interesting inputs.